Move factories into VMs
