+++ app/src/main/java/com/github/mik629/android/fundamentals/ui/movieslist/MoviesListViewModel.kt
@@ -3,11 +3,13 @@ package com.github.mik629.android.fundamentals.ui.movieslist
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.github.mik629.android.fundamentals.domain.model.Movie
 import com.github.mik629.android.fundamentals.domain.repositories.MoviesRepository
 import kotlinx.coroutines.launch
 import timber.log.Timber
+import javax.inject.Inject
 
 class MoviesListViewModel(
     private val moviesRepository: MoviesRepository
@@ -34,5 +36,13 @@ class MoviesListViewModel(
                 Timber.e(it)
             }
         }
+    }
+
+    class Factory @Inject constructor(
+        private val moviesRepository: MoviesRepository
+    ) : ViewModelProvider.Factory {
+
+        override fun <T : ViewModel?> create(modelClass: Class<T>): T =
+            MoviesListViewModel(moviesRepository) as T
     }
 }
